port to maven, and remove the old economy support. Everything goes through Vault.
