+++ .gitignore
@@ -1,9 +1,2 @@
-/nbproject/private/
-/.externalToolBuilders/
-/build/
-/dist/
-/bin
-/nbproject/pmd.settings
-/.idea/workspace.xml
-/.idea/ant.xml
-/.idea/workspace.xml
+target/
+.idea/
